Moved some functions into utilities module
