+++ icepack/utilities.py
@@ -14,6 +14,18 @@ import inspect
 import numpy as np
 import firedrake
 
+def facet_normal_2(mesh):
+    r"""Compute the horizontal component of the unit outward normal vector
+    to a mesh"""
+    ν = firedrake.FacetNormal(mesh)
+    return firedrake.as_vector((ν[0], ν[1]))
+
+
+def grad_2(q):
+    r"""Compute the horizontal gradient of a 3D field"""
+    return firedrake.as_tensor((q.dx(0), q.dx(1)))
+
+
 def diameter(mesh):
     r"""Compute the diameter of the mesh in the L-infinity metric"""
     X = mesh.coordinates.dat.data_ro
